Revert "drm: backport from 4.8-rc8"
This reverts commit 3941cd9ab424dab903c42fd16e116e4bb8e2c43f.

Conflicts:
	drivers/gpu/drm/i915/intel_ringbuffer.c

This is done since we want to align with following kernel:

https://android.googlesource.com/kernel/common.git/+/android-4.4-n

+++ drivers/gpu/drm/etnaviv/etnaviv_gpu.h
@@ -125,7 +125,7 @@ struct etnaviv_gpu {
 	u32 completed_fence;
 	u32 retired_fence;
 	wait_queue_head_t fence_event;
-	u64 fence_context;
+	unsigned int fence_context;
 	spinlock_t fence_spinlock;
 
 	/* worker for handling active-list retiring: */
